% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{any.nonascii}
\alias{any.nonascii}
\title{Check whether each string contains any ASCII characters}
\usage{
\method{any}{nonascii}(x, ...)
}
\arguments{
\item{x}{vector of strings}

\item{...}{additional arguments passed to grepl}
}
\value{
a logical vector the same length as x
}
\description{
Check whether each string contains any ASCII characters
}
