% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check.numeric}
\alias{check.numeric}
\title{Find and return values that look non-numeric}
\usage{
check.numeric(inp, val = TRUE, uvals = val, nsamp = NULL)
}
\arguments{
\item{inp}{An input vector (usually a character vector)}

\item{val}{passed to grep(value=val)}

\item{uvals}{Return only unique values?}

\item{nsamp}{The number of (down) samples to return.  For instances where inp is very large}
}
\description{
Find and return values that look non-numeric
}
