% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cluster_identical_rows}
\alias{cluster_identical_rows}
\title{Cluster identical rows where all values are identical}
\usage{
cluster_identical_rows(rcfilt, stopmax = Inf, verbose = FALSE)
}
\arguments{
\item{rcfilt}{the data frame to operate on}

\item{stopmax}{the number of pairs of rows to check if identical (it's not very efficient ...)}

\item{verbose}{verbose output}
}
\value{
a vector of cluster ID's, where all rows in a cluster are identical
}
\description{
Cluster identical rows where all values are identical
}
