% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{col2hex}
\alias{col2hex}
\title{Convert R color string and output the color as hex}
\usage{
col2hex(xx, transval = NULL)
}
\arguments{
\item{xx}{The R color string to convert}

\item{transval}{transparency (alpha) value, as 2-character hex string.}
}
\value{
A hex string representing the color and transparency
}
\description{
Convert R color string and output the color as hex
}
\examples{
col2hex("red", "2F")
}
