% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{expand_on}
\alias{expand_on}
\title{Split column values by a delimiter, then expand them into new rows}
\usage{
expand_on(indf, .var, delim = ";")
}
\arguments{
\item{indf}{input data.frame}

\item{.var}{the column in `indf` to split. should be character-type}

\item{delim}{the string delimiter to split on}
}
\value{
A data frame where data frae, `indf`, where the strings in  column, `.var`, have been split by delimiter, `delim`
}
\description{
Split column values by a delimiter, then expand them into new rows
}
\examples{
library(dplyr)
library(magrittr)
library(tidyr)
iris \%>\%
mutate_at(vars(Species), as.character) \%>\%
expand_on(Species, "o") \%>\%
select(Species) \%>\%
table
}
