% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{file.prefix}
\alias{file.prefix}
\alias{file.ext}
\title{Working with filenames}
\usage{
file.prefix(filename)

file.ext(filename)
}
\arguments{
\item{filename}{a string representing a filename}
}
\description{
Return the part of the file before the final period (i.e., the file extension).  The file does not have to exist
}
\section{Functions}{
\itemize{
\item \code{file.ext}: Return the part of the file before the final period (i.e., the file extension).  The file does not have to exist
}}

