% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{file_prefix}
\alias{file_prefix}
\alias{file_ext}
\title{Working with filenames}
\usage{
file_prefix(filename)

file_ext(filename)
}
\arguments{
\item{filename}{a string representing a filename}
}
\value{

}
\description{
Return the part of the file before the final period (i.e., the file extension).  The file does not have to exist
}
\section{Functions}{
\itemize{
\item \code{file_ext}: Return the part of the file before the final period (i.e., the file extension).  The file does not have to exist
}}

