% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fqtable}
\alias{fqtable}
\title{Tabulate and return relative frequencies of values}
\usage{
fqtable(vect, scaling = 1, hightolow = TRUE, ...)
}
\arguments{
\item{vect}{input vector}

\item{scaling}{multiple frequencies by this scaling factor (e.g, 100 to get percentages)}

\item{hightolow}{should returned frequency table elements be sorted from high to low?}

\item{...}{additional arguments passed to `table` function}
}
\value{

}
\description{
Tabulate and return relative frequencies of values
}
