% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gen7singles2018}
\alias{gen7singles2018}
\title{GloVe Model of Pokémon}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 181 rows and 66 columns.
}
\usage{
gen7singles2018
}
\description{
A GloVe model that trained on
11,446 Pokémon parties (for single battles in 2018),
while considering each party as a sentence
and each Pokémon as a word, and pruning vocabularies.
}
\keyword{datasets}
