% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{get.duplicated}
\alias{get.duplicated}
\title{Identify duplicate values within a vector}
\usage{
get.duplicated(vect)
}
\arguments{
\item{vect}{A vector that potentially contains duplicates}
}
\value{
the names of duplicated values
}
\description{
Identify duplicate values within a vector
}
