% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{get_anno}
\alias{get_anno}
\title{Extract (generally) parenthetical annotations from within a string}
\usage{
get_anno(
  x,
  opening = "(",
  closing = ")",
  auto.fix = FALSE,
  other.closing = c("}", "\\\\]", "\\\\)")
)
}
\arguments{
\item{x}{string to extract annotations from}

\item{opening}{character that indicates the start of an annotation block}

\item{closing}{character that indicates the end of an annotation block}

\item{auto.fix}{if TRUE, then additional closing characters are checked when an open block has no closing character.}

\item{other.closing}{other closing characters to look for.  only used when auto.fix=TRUE}
}
\value{

}
\description{
Extract (generally) parenthetical annotations from within a string
}
\examples{
 get_anno("this is a (test)")
}
