% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{get_duplicated}
\alias{get_duplicated}
\alias{has_duplicated}
\alias{which.duplicated}
\title{List duplicate values}
\usage{
get_duplicated(vect)

has_duplicated(...)

which.duplicated(vect)
}
\arguments{
\item{vect}{A vector that potentially contains duplicates}

\item{...}{arguments passed to `get_duplicated`}
}
\value{
the names of duplicated values
}
\description{
List duplicate values
}
\section{Functions}{
\itemize{
\item \code{has_duplicated}: check if any are duplicated

\item \code{which.duplicated}: return indices of all values that are not unique in `vect`
}}

\examples{
get_duplicated( c(1:10, 1:5) )
get_duplicated( c(1:10) )
}
