% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{has_nonascii}
\alias{has_nonascii}
\alias{any_nonascii}
\title{Check for any non-ASCII characters}
\usage{
has_nonascii(x, ...)

any_nonascii(...)
}
\arguments{
\item{x}{vector of strings}

\item{...}{additional arguments passed to grepl}
}
\value{
a logical vector the same length as x
}
\description{
Check for any non-ASCII characters
}
\section{Functions}{
\itemize{
\item \code{any_nonascii}: (DEPRECATED) alias for has_nonascii, included for back-compatibility
}}

