% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_preserve.R
\name{load_preserve}
\alias{load_preserve}
\title{Wrapper around base::load that preserves the value of any user-specified variables}
\usage{
load_preserve(rda.path, ...)
}
\arguments{
\item{rda.path}{path to loadable file, passed to `base::load` via the `file` argument}

\item{...}{strings or a character vector of object names in current envir that should be restored after the load}
}
\value{
nothing
}
\description{
Wrapper around base::load that preserves the value of any user-specified variables
}
\examples{
#load_preserve("path_to_RData_file", "DO_WRITES")
#load_preserve("path_to_RData_file", c("adf.wide","batch_names","load_exclude","condition_groups","et.modify"))

}
