% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{low.incidence}
\alias{low.incidence}
\alias{which.low}
\title{Return logical vector indicating location of low-incidence values in, x}
\usage{
low.incidence(x, min.count = 1, min.freq, use.median = FALSE)

which.low(...)
}
\arguments{
\item{x}{Input vector}

\item{min.count}{the threshold number of occurrences}

\item{min.freq}{if set, then use frequency threshold instead of count}

\item{use.median}{Set minimum count to the median occurence in x}
}
\description{
Return logical vector indicating location of low-incidence values in, x
}
\section{Functions}{
\itemize{
\item \code{which.low}: return the indices of low incidence values
}}

