% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{low_incidence}
\alias{low_incidence}
\alias{are_few}
\alias{which_low}
\title{Return logical vector indicating location of low-incidence values in, x}
\usage{
low_incidence(x, min.count = 1, min.freq, use.median = FALSE)

are_few(...)

which_low(...)
}
\arguments{
\item{x}{Input vector}

\item{min.count}{the threshold number of occurrences}

\item{min.freq}{if set, then use frequency threshold instead of count}

\item{use.median}{Set minimum count to the median occurence in x}

\item{...}{arguments passed to `low_incidence`}
}
\value{

}
\description{
Return logical vector indicating location of low-incidence values in, x
}
\section{Functions}{
\itemize{
\item \code{are_few}: an alias for `low_incidence`

\item \code{which_low}: return the indices of low incidence values
}}

\examples{
tmprnd = round(iris$Sepal.Length)
table(tmprnd, low_incidence(tmprnd, min.count = 5))
}
