% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{n.unique}
\alias{n.unique}
\title{Return unique values}
\usage{
n.unique(x)
}
\arguments{
\item{x}{A vector}
}
\value{

}
\description{
Function that returns the number of unique values in a particular vector
}
