% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{ngt_tokenize}
\alias{ngt_tokenize}
\title{Function that chops a strings into tokens based on one or two delimiters}
\usage{
ngt_tokenize(str, first.lev = ",", second.lev = NA)
}
\arguments{
\item{str}{A string}

\item{first.lev}{Delimiter to chop string on first}

\item{second.lev}{Delimiter to chop up results from first one}
}
\value{

}
\description{
Function that chops a strings into tokens based on one or two delimiters
}
