% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NgtIndex.R
\name{NgtIndex}
\alias{NgtIndex}
\title{R6 Class for Graph and Tree Based Index}
\description{
Graph and tree based index.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{path of the index.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NgtIndex-new}{\code{NgtIndex$new()}}
\item \href{#method-NgtIndex-open}{\code{NgtIndex$open()}}
\item \href{#method-NgtIndex-create}{\code{NgtIndex$create()}}
\item \href{#method-NgtIndex-search}{\code{NgtIndex$search()}}
\item \href{#method-NgtIndex-save}{\code{NgtIndex$save()}}
\item \href{#method-NgtIndex-remove}{\code{NgtIndex$remove()}}
\item \href{#method-NgtIndex-refine_anng}{\code{NgtIndex$refine_anng()}}
\item \href{#method-NgtIndex-get_object}{\code{NgtIndex$get_object()}}
\item \href{#method-NgtIndex-reset_defaults}{\code{NgtIndex$reset_defaults()}}
\item \href{#method-NgtIndex-get_info}{\code{NgtIndex$get_info()}}
\item \href{#method-NgtIndex-build_index}{\code{NgtIndex$build_index()}}
\item \href{#method-NgtIndex-batch_insert}{\code{NgtIndex$batch_insert()}}
\item \href{#method-NgtIndex-insert}{\code{NgtIndex$insert()}}
\item \href{#method-NgtIndex-export_index}{\code{NgtIndex$export_index()}}
\item \href{#method-NgtIndex-import_index}{\code{NgtIndex$import_index()}}
\item \href{#method-NgtIndex-close}{\code{NgtIndex$close()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-new"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{NgtIndex} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$new(path, sub_dir = rand_name("NgtIndex"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to the NGT index.}

\item{\code{sub_dir}}{sub-directory of the index.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{NgtIndex} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-open"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-open}{}}}
\subsection{Method \code{open()}}{
Opens a NGT index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$open(read_only = FALSE, tree_disabled = FALSE, log_disabled = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{read_only}}{whether the index is read only.}

\item{\code{tree_disabled}}{whether the tree is disabled.}

\item{\code{log_disabled}}{whether the log is disabled.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-create"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-create}{}}}
\subsection{Method \code{create()}}{
Creates an empty index with the specified parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$create(
  dimension,
  edge_size_for_creation = 10,
  edge_size_for_search = 40,
  distance_type = c("l2", "l1", "normalized_l2", "hamming", "jaccard", "sparse_jaccard",
    "angle", "normalized_angle", "cosine", "normalized_cosine", "normalized_l2"),
  object_type = c("float", "byte", "float16")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dimension}}{dimension of the vectors.}

\item{\code{edge_size_for_creation}}{number of edges for each node in the graph.}

\item{\code{edge_size_for_search}}{number of edges to search.}

\item{\code{distance_type}}{distance type.}

\item{\code{object_type}}{object type.}

\item{\code{sub_dir}}{sub directory to store the index in.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-search"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-search}{}}}
\subsection{Method \code{search()}}{
Searches for the k approximate nearest neighbors
of the specifiecd query object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$search(
  query,
  k = 20L,
  epsilon = 0.1,
  edge_size = -1L,
  expected_accuracy = -1,
  with_distance = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{query object.}

\item{\code{k}}{number of nearest neighbors.}

\item{\code{epsilon}}{epsilon which defines the explored range for the graph}

\item{\code{edge_size}}{number of edges for each node to explore the graph}

\item{\code{expected_accuracy}}{expected accuracy.}

\item{\code{with_distance}}{whether to return distance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-save"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-save}{}}}
\subsection{Method \code{save()}}{
Saves the index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$save(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to save the index. defaults to the path of the index.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-remove"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-remove}{}}}
\subsection{Method \code{remove()}}{
Removes objects from the index by their IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$remove(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{IDs of the objects to be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
integers; \code{ids} is returned invisibly as is.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-refine_anng"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-refine_anng}{}}}
\subsection{Method \code{refine_anng()}}{
Refines the index with the specified parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$refine_anng(
  epsilon,
  accuracy,
  num_edges,
  num_edges_for_search,
  batch_size
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{epsilon}}{epsilon which defines the explored range for the graph}

\item{\code{accuracy}}{expected accuracy.}

\item{\code{num_edges}}{number of edges for each node to explore the graph}

\item{\code{num_edges_for_search}}{number of edges to search.}

\item{\code{batch_size}}{batch size.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-get_object"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-get_object}{}}}
\subsection{Method \code{get_object()}}{
Gets objects from the index by their IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$get_object(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{IDs of the objects to be retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-reset_defaults"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-reset_defaults}{}}}
\subsection{Method \code{reset_defaults()}}{
Resets the default parameters of the index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$reset_defaults(
  num_of_search_objects,
  search_radius,
  epsilon,
  edge_size,
  expected_accuracy
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_of_search_objects}}{number of search objects.}

\item{\code{search_radius}}{search radius.}

\item{\code{epsilon}}{epsilon.}

\item{\code{edge_size}}{edge size.}

\item{\code{expected_accuracy}}{expected accuracy.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-get_info}{}}}
\subsection{Method \code{get_info()}}{
Gets information of the index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$get_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-build_index"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-build_index}{}}}
\subsection{Method \code{build_index()}}{
Builds the search index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$build_index(num_threads = 1L, target_size_of_graph = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_threads}}{number of threads
to be used for building a search index.}

\item{\code{target_size_of_graph}}{target size of the graph.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-batch_insert"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-batch_insert}{}}}
\subsection{Method \code{batch_insert()}}{
Inserts data into the index and build a search index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$batch_insert(data, num_threads = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data to be inserted.}

\item{\code{num_threads}}{number of threads to be used for building the index.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
integers; the IDs of the inserted objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-insert"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-insert}{}}}
\subsection{Method \code{insert()}}{
Inserts a vector to the index.
To search with the index,
you need to call \code{build_index} after call this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$insert(vec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{vector to be inserted.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
integers; the IDs of the inserted objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-export_index"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-export_index}{}}}
\subsection{Method \code{export_index()}}{
Exports an index to a file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$export_index(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to save the index.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-import_index"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-import_index}{}}}
\subsection{Method \code{import_index()}}{
Imports an index from a file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$import_index(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to load the index.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NgtIndex-close"></a>}}
\if{latex}{\out{\hypertarget{method-NgtIndex-close}{}}}
\subsection{Method \code{close()}}{
Closes the index.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NgtIndex$close()}\if{html}{\out{</div>}}
}

}
}
