% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{number2binary}
\alias{number2binary}
\title{Convert number to bit vector}
\usage{
number2binary(number, noBits)
}
\arguments{
\item{number}{a number to convert to a binary vector}

\item{noBits}{the final length of the binary vector.  If missing, return the full (native) bit vector}
}
\value{
a vector of 0's and 1's
}
\description{
Convert number to bit vector
}
\examples{
number2binary(1)
number2binary(2)
number2binary(3)
number2binary(4)
}
