% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rand_name}
\alias{rand_name}
\title{Make a random name}
\usage{
rand_name(prefix = "index", len = 5)
}
\arguments{
\item{prefix}{A single character string}

\item{len}{An integer for the number of random characters}
}
\value{
A character string with the prefix and random letters separated by
and underscore.
}
\description{
Make a random name
}
\keyword{internal}
