% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.R
\name{read.cb}
\alias{read.cb}
\title{Read data table from system clipboard}
\usage{
read.cb(...)
}
\arguments{
\item{...}{arguments passed to read.table}
}
\value{
a data frame created from clipboard data
}
\description{
Read data table from system clipboard
}
