% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{reducews}
\alias{reducews}
\alias{rm.paren}
\alias{rm.brack}
\alias{rm.curly}
\alias{remove.blocks}
\title{Reduce consecutive, internal whitespace characters (not leading or trailing) to a single whitespace character}
\usage{
reducews(x)

rm.paren(x, ...)

rm.brack(x, ...)

rm.curly(x, ...)

remove.blocks(x, ...)
}
\arguments{
\item{x}{A chacter vector}

\item{...}{additional arguments passed to `regex` function}
}
\value{

}
\description{
Reduce consecutive, internal whitespace characters (not leading or trailing) to a single whitespace character
}
\section{Functions}{
\itemize{
\item \code{rm.paren}: remove text within parentheses

\item \code{rm.brack}: remove text within square braces

\item \code{rm.curly}: remove text within curly braces

\item \code{remove.blocks}: remove text within parentheses, or square or curly braces
}}

\examples{
reducews("CCN      51")
}
\seealso{
[stringr:::str_squish()] which does essentially the same thing
}
