% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{unfactor}
\alias{unfactor}
\title{Convert all factors in a data.frame to strings}
\usage{
unfactor(df)
}
\arguments{
\item{df}{the data frame to operate on}
}
\description{
Convert all factors in a data.frame to strings
}
