% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{vec2range}
\alias{vec2range}
\title{Convert a vector of integers in to a matrix of ranges}
\usage{
vec2range(vec, with.max = NULL)
}
\arguments{
\item{vec}{A vector of integers}

\item{with.max}{(optional) A number to use as the end point of the last range}
}
\value{

}
\description{
A function which converts a one-dimensional vector into a 2-column matrix of pairwise ranges that represent the range of numbers between subsequent values
}
