% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{which.top}
\alias{which.top}
\alias{which.na}
\alias{which.bottom}
\title{Expanded which() functions}
\usage{
which.top(vect, n = 5)

which.na(vect)

which.bottom(vect, n = 5)
}
\arguments{
\item{vect}{a vector}

\item{n}{number of elements to return}
}
\value{
Returns the positions of top or bottom `n` elements of a vector
}
\description{
Function that returns the indices of the top N observations
}
\section{Functions}{
\itemize{
\item \code{which.na}: return the indices of NA values

\item \code{which.bottom}: return the indices of the N smallest elements
}}

\examples{
which.top(1:10, 7)
which.bottom(1:10, 3)
}
