% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{which.topn}
\alias{which.topn}
\alias{which.na}
\alias{which.top5}
\alias{which.top10}
\alias{which.botn}
\alias{which.bot5}
\alias{which.bot10}
\title{Expanded which() functions}
\usage{
which.topn(vect, n)

which.na(vect)

which.top5(...)

which.top10(...)

which.botn(vect, n)

which.bot5(...)

which.bot10(...)
}
\arguments{
\item{vect}{A numeric vector}

\item{n}{}
}
\description{
Function that returns the indices of the top N observations
}
\section{Functions}{
\itemize{
\item \code{which.na}: return the indices of NA values

\item \code{which.top5}: return the indices of the top 5 largest elements

\item \code{which.top10}: return the indices of the top 10 largest elements

\item \code{which.botn}: return the indices of the N smallest elements

\item \code{which.bot5}: return the bottom 5 element

\item \code{which.bot10}: return the bottom 10 elements
}}

