% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fqtable}
\alias{fqtable}
\title{Calculate table and return relative frequencies}
\usage{
fqtable(vect, scaling = 1, ...)
}
\arguments{
\item{scaling}{multiple frequencies by this scaling factor (e.g, 100 to get percentages)}

\item{...}{}
}
\description{
Calculate table and return relative frequencies
}
