% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{reducews}
\alias{reducews}
\title{Reduce consecutive, internal whitespace characters (not leading or trailing) to a single whitespace character}
\usage{
reducews(x)
}
\arguments{
\item{x}{A chacter vector}
}
\description{
Reduce consecutive, internal whitespace characters (not leading or trailing) to a single whitespace character
}
\examples{
reducews("CCN      51")
}
\seealso{
[stringr:::str_squish()] which does essentially the same thing
}
